Add files via upload
SpeedBump - A Hardware Project For The Tandy Radio Shack TRS-80 Color Computer 3 by "Ankle" Bob. This is a relatively simple project to restore, to an extent, the Color Computer 1 and 2 high-speed POKE to the Tandy Color Computer 3. Some might consider this to be a bit of folly, after all, "Doesn't the Color Cumputer 3 possess a superior high-speed POKE?" Yes, indeed it does. However, one of the very first things that I noticed about my newly acquired CoCo 3, circa 1986, was the fact that the old high-speed POKE no longer functioned. This had the effect of rendering some of my favorite programs, such as "Paint King", unusable, or nearly so. Now, it was a very simple matter to patch "Paint King" and similar BASIC and unprotected programs by simply substituting the new POKE for the old. In fact, this actually improved the functionality of programs such as "Paint King". This was due to the CoCo 3 high-speed POKE NOT BEING ADDRESS DEPENDENT, as it was on the CoCo 1 and 2. Many machine-language and "protected" programs proved difficult, and in some cases, impossible to patch. After wasting many hours with a "Disk Zapper", hacking and patching, I set about to solve the problem, once and for all, in hardware. In theory, at least, it should be quite possible and simple enough to socket the MC68B09E CPU and stick a circuit between the CPU and the rest of the system to monitor the Address Buss and pass all addresses straight through, EXCEPT for the old high/slow-speed POKE addresses. The old high/slow-speed POKE addresses would be modified (converted to the NEW high/slow-speed addresses) and passed through as the new addresses. Thanks to the amazingly versatile (at the time) PALs and PLDs of the era, this is exactly what I did. Two PLD/PALs provided an elegant two chip solution. A request from my nephew, "Little John", prompted me, in part, to redesign the device utilizing a single 22V10 GAL. This was a desirable decision due to the scarcity of the older PLD and PAL devices and the fact that often, when purchased, the older devices have already been programmed. Being fusible link based devices, this renders them unusable and completely useless. A GAL chip is also much faster and flash-based (reusable) with some offering In System Programming (such as the ISPGAL22V10 from Lattice.) The ISP version is not used here, as they are a bit difficult to obtain, so you'll need a GAL programmer to complete this project, or at least access to one. The g22v10d is reccommended due to its higher speed than the plain vanilla g22v10, although it should work as well. The A,B and C versions could be used as well. The CUPL code is self-explanatory and this truly is a very simple device. Note that the two old addresses are simply converted to the two new addresses. This effectively restores the old high-speed POKE with the exception that the machine will run at 1.79MHz at all times when the POKE is enabled. This is in contrast to the CoCo 1 and 2, which only ran at 1.79MHz during access to the ROM area ($8000 - $FEFF) and only if the ROMs were enabled and also at a small section of I/O addresses. It ran at .89MHz for accesses to the lower 32K (RAM area) and, if the machine were a 64K RAM machine running in All RAM mode, it would stay at .89MHz during accesses to the $8000-$FEFF area as well. In short, the old high speed POKE only functioned if the BASIC ROMs were in circuit and being executed by the CPU (using the original 74LS783/MC6883 S.A.M. - later versions of the MC6883 and the 74LS785 allowed the address dependent mode to function while in "ALL-RAM" mode as well). This version, the CC3-SpeedBump-V1R1-30-MARCH-1987 is licensed under a Creative Commons Attribution-ShareAlike 4.0 International (CC BY-SA 4.0) License: https://creativecommons.org/licenses/by-sa/4.0/